requirements.txt
discord.py
playwright
matplotlib
python-dotenv
requests
